split up listitem to make it easier
