+++ src/components/sdm/ListItem/ListItemHeader.js
@@ -1,0 +1,6 @@
+import {Observable as $} from 'rx'
+
+import {ListItem} from './ListItem'
+
+export const ListItemHeader = sources =>
+  ListItem({...sources, classes$: $.just({header: true})})
