+++ src/components/sdm/ListItem/ListItemWithMenu.js
@@ -1,0 +1,35 @@
+import {Observable as $} from 'rx'
+import combineLatestObj from 'rx-combine-latest-obj'
+
+import {div} from 'cycle-snabbdom'
+
+import {ListItemClickable} from './ListItemClickable'
+import {Menu} from 'components/sdm'
+
+export const ListItemWithMenu = sources => {
+  const item = ListItemClickable(sources)
+
+  const isOpen$ = item.click$.map(true).startWith(false)
+
+  const children$ = sources.menuItems$ || $.just([])
+
+  const menu = Menu({
+    ...sources,
+    isOpen$,
+    children$,
+  })
+
+  const viewState = {
+    itemDOM$: item.DOM,
+    menuDOM$: menu.DOM,
+  }
+
+  const DOM = combineLatestObj(viewState)
+    .map(({itemDOM, menuDOM}) =>
+      div({},[itemDOM, menuDOM])
+    )
+
+  return {
+    DOM,
+  }
+}
