+++ src/components/sdm/index.js
@@ -1,19 +1,15 @@
 require('./styles.scss')
 
-export {Toolbar, AccentToolbar} from './Toolbar'
-export {Menu, MenuItem} from './Menu'
+export * from './Card'
+export * from './List'
+export * from './ListItem'
 
-export {
-  Avatar,
-  MediumAvatar,
-  LargeAvatar,
-} from './Avatar'
+export * from './Toolbar'
+export * from './Menu'
 
-export {
-  RaisedButton,
-  FlatButton,
-  OkAndCancel,
-} from './Button'
+export * from './Avatar'
+
+export * from './Button'
 
 export {Fab} from './Fab'
 
@@ -23,25 +19,5 @@ export {ToggleControl} from './ToggleControl'
 export {TextAreaControl} from './TextAreaControl'
 export {CheckboxControl} from './CheckboxControl'
 
-export * from './List'
+export * from './Dialog'
 
-export {
-  Dialog,
-  BaseDialog,
-} from './Dialog'
-
-export {
-  ListItem,
-  ListItemClickable,
-  ListItemToggle,
-  ListItemCheckbox,
-  ListItemWithMenu,
-  ListItemNavigating,
-  ListItemWithDialog,
-  ListItemCollapsible,
-  ListItemCollapsibleTextArea,
-  ListItemHeader,
-  ListItemTextArea,
-} from './ListItem'
-
-export * from './Card'
